back/requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
torch
python-dotenv